Minor fix for the recommended startup.
The old setting doesn't work when Vim reads a file from stdin, f.i.:

    diff file1 file2 | vim -

Then `argc() == 0`, since no file has been named on the command line,
but we don't want to open NERDTree, as we already have a file to play
with.  However, when Vim raeds a file from stdin the buffer is still
marked as modified.  Hence the patch.