WIP: Allow users to disallow mappings of some keys
Per the discussion in #255, it'd be nice to tell NERDTree not to remap
certain keys, while retaining the standard NERDTree functionality for
other keys (like `p`).

This works by checking a list the user can specify and not mapping keys
in that list.

I've kept most of the standard NERDTree mappings, but kept the `<c-k>`
and `<c-j>` mappings from my dotfiles by adding: >
    let g:NERDNoOverride=['<lt>C-j>', '<lt>C-k>']
<
To my vimrc.

Note: this is a work in progress, the following issues still exist:
 - The `g:NERDNoOverride` variable is case sensitive
  - I'm new to vim scripting, so I'm not sure how to fix this/if it
    matters yet
 - The strings in `g:NERDNoOverride` are case sensitive and must be
   specified exactly as they're seen in the NERDTree plugin
 - Keys like '<c-k>' must be specified as '<lt>C-k>' for the time being.
   I think NERDTree converts these internally, so that'll need to be
   fixed as well.