Added 'NERDTreeReuseWindows' option to choose whether to reuse windows while opening files
New option name: NERDTreeReuseWindows
Values: 0 or 1
Default: 1

While working with several windows and/or tabs there may be situation when
you ask NERDTree to open file already being edited in other window.
Default behavior is to jump to that window. If this option is not set
NERDTree will always open selected file in previous window.