Add an option to hide or display trailing '*' on executable files.
If the option NERDTreeShowExecutableFlag is set to 0 in the .vimrc, all
executable files in NERDTree will be shown without the trailing
asterisk. By default, or if the value is set to 1, the files will be
displayed with the asterisk.

The usual highlighting associated with executable files is broken when
this value is 0, because the syntax matching expects the trailing
asterisk to be there. I know no way around this, but I believe it's a
small price to pay for the new functionality.