Display file size with commas as thousands separators.
This change will display the file size in an easier to read format, and
still display its exact value. It uses the stat command (with different
switches and formats between Unix and OSX) to get the file size, sed to
add the commas to it, and sed again to replace the original size with
the modified number.