Implements #237. Make split- and tab-open reuse windows.
Instead of adding a new option or command to toggle the behavior, as
shown in #237, this commit makes the behavior the same as for the `o`
command for file nodes. If the file is already open in a window, it is
reused, instead of being opened in a new split or tab.

Since this is a change of behavior, splitting a window to duplicate
the buffer already in it will need to be done with the vim command
`:sp`, `:vsp`, or `:tabe %`.